+++ HISTORY.md
@@ -7,6 +7,7 @@
 * threads: optimized frequency distribution analysis (uniques, distribution) for lists of colors
 * objects: tweaked LISP code for new color primitives
 * objects: updated new primitives list for hiding new blocks in old puzzles
+* blocks, objects, threads: new %color selector
 
 ### 2025-04-03
 * tables: further rendering speed-up